Back to old versioning
Previously developers weren't able to figure out what version of the library is used in their projects if they're working in a team. That means they don't know what methods the integrated version of the script supports. They're also not able to find out if the new version of the script is loaded properly if the site is cached on the server- or client-side.